Deacidify usernotify
